DC-2978 [Account Color] Account color should be changed directly in sift/draft folder
