+++ app/internal_packages/account-color/lib/main.es6
@@ -1,0 +1,3 @@
+export function activate() {}
+
+export function deactivate() {}
