+++ app/internal_packages/draft-list/lib/draft-list-columns.jsx
@@ -2,6 +2,7 @@
 import React from 'react';
 import { Utils, AccountStore, FocusedPerspectiveStore } from 'mailspring-exports';
 import { InjectedComponentSet, ListTabular, LabelColorizer } from 'mailspring-component-kit';
+import AccountColor from '../../account-color/lib/account-color';
 
 function snippet(html) {
   if (!(html && typeof html === 'string')) {
@@ -21,29 +22,6 @@ function subject(subj) {
   return Utils.extractTextFromHtml(subj);
 }
 
-function renderAccountColor(draft) {
-  if (!AppEnv.config.get('core.appearance.showAccountColor')) {
-    return null;
-  }
-  const current = FocusedPerspectiveStore.current();
-  if (current.accountIds.length <= 1) {
-    return null;
-  }
-  const accounts = AccountStore.accounts().map(account => account.id);
-  const accountId = draft.accountId;
-  const account = AccountStore.accountForId(accountId);
-  const color =
-    account.color !== undefined
-      ? account.color
-      : accounts.findIndex(account => account === accountId) + 1;
-  return (
-    <div
-      className={`account-color`}
-      style={{ background: LabelColorizer.accountColors()[color] }}
-    ></div>
-  );
-}
-
 const ParticipantsColumn = new ListTabular.Column({
   name: 'Participants',
   width: 180,
@@ -56,12 +34,15 @@ const ParticipantsColumn = new ListTabular.Column({
           <div className="participants-inner">
             <span>{list.map(p => p.displayName()).join(', ')}</span>
           </div>
-          {renderAccountColor(draft)}
+          <AccountColor message={draft} />
         </div>
       );
     } else {
       return (
-        <div className="participants no-recipients">(No Recipients){renderAccountColor(draft)}</div>
+        <div className="participants no-recipients">
+          (No Recipients)
+          <AccountColor message={draft} />
+        </div>
       );
     }
   },
