+++ app/internal_packages/sift-list/lib/sift-list-columns.jsx
@@ -9,6 +9,8 @@ import {
 } from 'mailspring-exports';
 import { ListTabular, InjectedComponent, LabelColorizer } from 'mailspring-component-kit';
 import SiftQuickActions from './sift-quick-actions';
+import AccountColor from '../../account-color/lib/account-color';
+
 function snippet(html) {
   if (!(html && typeof html === 'string')) {
     return '';
@@ -26,29 +28,6 @@ function subject(subj) {
   }
   return Utils.extractTextFromHtml(subj);
 }
-
-const renderAccountColor = message => {
-  if (!AppEnv.config.get('core.appearance.showAccountColor')) {
-    return null;
-  }
-  const current = FocusedPerspectiveStore.current();
-  if (current.accountIds.length <= 1) {
-    return null;
-  }
-  const accounts = AccountStore.accounts().map(account => account.id);
-  const accountId = message.accountId;
-  const account = AccountStore.accountForId(accountId);
-  const color =
-    account.color !== undefined
-      ? account.color
-      : accounts.findIndex(account => account === accountId) + 1;
-  return (
-    <div
-      className={`account-color`}
-      style={{ background: LabelColorizer.accountColors()[color] }}
-    ></div>
-  );
-};
 
 const participants = message => {
   let isSent = false;
@@ -69,12 +48,15 @@ const participants = message => {
         <div className="participants-inner">
           <span>{list.map(p => p.displayName()).join(', ')}</span>
         </div>
-        {renderAccountColor(message)}
+        <AccountColor message={message} />
       </div>
     );
   } else {
     return (
-      <div className="participants no-recipients">(No Recipients){renderAccountColor(message)}</div>
+      <div className="participants no-recipients">
+        (No Recipients)
+        <AccountColor message={message} />
+      </div>
     );
   }
 };
